Cub3D Fonctionnel / Mur violets

+++ cub3d/srcs/cub3d_utils.c
@@ -6,7 +6,7 @@
 /*   By: mbosson <marvin@le-101.fr>                 +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2020/01/15 14:53:31 by mbosson      #+#   ##    ##    #+#       */
-/*   Updated: 2020/01/28 10:45:55 by mbosson     ###    #+. /#+    ###.fr     */
+/*   Updated: 2020/02/04 16:17:20 by mbosson     ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
@@ -62,7 +62,7 @@ int tablen(char **tab)
 	return (i);
 }
 
-long	whose_higher(r_list raycasting)
+double	whose_higher(r_list raycasting)
 {
 	if (raycasting.long_horizontal < raycasting.long_vertical
 			&& raycasting.long_horizontal > 0)
